Added SQLite functionality via service
Added a service to handle SQLite interactions, and created a factory for
the Team table. Added Team table creation script to app.js. Updated
pit.html to use the value passed in by PitCtrl. Updated the PitCtrl
controller to use the service to retrieve team names and numbers from
the database. Added some commented-out bootstrap and cleanup code for
pre-loading the database and emptying the database.