Added cascading select to Pit Scouting page
Added Robot factory to Services file. Fixed issue with the return value
from the Get method for the Team factory (it needs to return all; the
restriction by ID is handled in the SQL). Added functionality to the Pit
page and corresponding controller to change the values in Robot based on
the value selected from the Team field.

Added some bootstrap code to app.js to create the Robot table and to
populate the tables with staged data.