Updated the Readme file
Updated the readme file to reflect the change in direction the project
is taking regarding the transferring of data. No longer will the app
support an ad-hoc network for sharing data, as this would violate the
FRC rules regarding wifi hotspots. The app will now use flash media to
store files for transferring between mobile devices and the server.