Updated to retrieve match numbers from database
Updated App.js to add more bootstrap code to create the match table and
populate it with test data. Updated Controllers.js to pull match data
from the database. Deleted content from the selectRobotMatch method
to be filled in later with SQLite calls. Updated the match.html to
populate the Team, Robot, and Match fields from the database. Created
factory code in Services.js for matches.