Set fields to retrieve from and persist to DB
In App.js, updated the Robot table to include all of the fields that
are captured on the Pit page. Added additional an bootstrap data
store insert to have two robots for one team. Added the Match table.

In Controllers.js, removed the values stored for a robot from the
$scope.data and moved them into $scope.robot.*. Updated the
selectTeam function so that it only retrieves a list of robots for
the team. Added a selectRobot function that retrieves values for a
robot from the data store. This function is a work in progress.
Started working on the robotChanged method for persisting values
selected in the form to the data store.

In Services.js, in the core factory, changed the return back to
DBA.getById when a single value is being returned. Renamed the get
method in the Robot factory to getByTeam to differentiate it from the
new getById function. Added an update method to persist robot pit
values to the data store.

In Pit.html, updated ng-change values to match changes in the
controller. Updated ng-model to no longer reference $scope.data.* but
instead $scope.*. Changed the name of the Robot field from robot to
selectedRobot to more easily distinguish it in the controller from
the robot object that is being retrieved/updated. (Naming things is
hard! http://martinfowler.com/bliki/TwoHardThings.html)

Note that I try to commit code only when everything is more-or-less
working, but today is an FRC workday, and I need to get the latest
version of the app onto some tablets for testing with the students!