Worked on the data layer for cascade selects
In App.js, added a drop table statement for the new table, RobotMatch.
Reformatted the create table statement for the Robot table to make it
easier to read. Added a create table statement for RobotMatch table.
Added a second robot to the Robot table in the bootstrap code, and a
record for RobotMatch.

In Controllers.js, added code to reset the robot field when a new team
value is selected. Updated the MatchScoutingController to add
RobotMatch factory to it. Removed the $scope.data from the reset
method, as it ended up not being helpful. Removed all but the values
used to populate the fields from $scope.data. Updated
$scope.selectTeam so that it preserves the selected team and match,
while resetting the other fields. Added $scope.selectRobot to load
values from the data store when both a robot and match are selected,
or set them all to unselected values if there is no record for the
selected robot. Added $scope.selectMatch, which just called
$scope.selectRobot. Added $scope.robotChanged to update all fields
when a different robot is selected.

In Services.js, updated self.getById to check for a value before
returning it. Reformatted its DBA.query for readability. Updated
Robot's self.update for readability.

In Match.html, updated some parameters to match changes made in the
controller code.