Added Android platform, whitelist plugin
Added the Android platform to the project (though I thought I had
already done this). I also added the Cordova whitelist plugin, which is
required in later version of Ionic. In my previous commit, I added some
code to www/index.html that is required to work with the whitelist
plugin.
E