Last-minute changes prior to Palmetto. Moved the database table create/delete functionality out of App.js and into Controller.js. Added buttons for those functions on the Settings page.
Also separated out the generate-sample-data functionality into
separate buttons since chaining wasn't working well. The user needs
to wait for each of the Toast messages to appear indicating that the
database record-generation process has completed, before moving onto
the next button. I'm still trying to figure out if the last blue
button needs to be clicked... preliminary testing seems to indicate
it does not.

Combined the MatchCtrl and MatchController controllers into one.
I plan to do the same thing for PitCtrl and PitController.

Still left to be done is to add functionality for the plus/minus
buttons.