Resolved data export issue
Resolved the issue with the exported data containing the database id
for the team record, rather than the number of the team.

Also made some cosmetic changes, including new text for the Home page,
new text for the Sync page, and warning text on the Settings page.

Changed the labels on some components on the Match page so that they
don't get cropped.

Updated the code that sets default values in fields when a new team/
robot/match combination is selected so that the drop-downs all have a
default of "Unknown."